refactor(channel): extract contract methods to a separate class
