+++ src/channel/Base.ts
@@ -1,6 +1,6 @@
 /*
  * ISC License (ISC)
- * Copyright (c) 2018 aeternity developers
+ * Copyright (c) 2022 aeternity developers
  *
  *  Permission to use, copy, modify, and/or distribute this software for any
  *  purpose with or without fee is hereby granted, provided that the above
@@ -19,7 +19,7 @@ import { EventEmitter } from 'events';
 import { w3cwebsocket as W3CWebSocket } from 'websocket';
 import { snakeToPascal } from '../utils/string';
 import { buildTx, unpackTx } from '../tx/builder';
-import { MIN_GAS_PRICE, Tag } from '../tx/builder/constants';
+import { Tag } from '../tx/builder/constants';
 import * as handlers from './handlers';
 import {
   initialize,
@@ -37,7 +37,6 @@ import {
 } from './internal';
 import { ChannelError } from '../utils/errors';
 import { Encoded } from '../utils/encoder';
-import { ContractCallReturnType } from '../apis/node';
 import { pause } from '../utils/other';
 
 function snakeToPascalObjKeys<Type>(obj: object): Type {
@@ -48,30 +47,6 @@ function snakeToPascalObjKeys<Type>(obj: object): Type {
 }
 
 type EventCallback = (...args: any[]) => void;
-
-interface CallContractOptions {
-  amount?: number | BigNumber;
-  callData?: Encoded.ContractBytearray;
-  abiVersion?: number;
-  contract?: Encoded.ContractAddress;
-  returnValue?: any;
-  gasUsed?: number | BigNumber;
-  gasPrice?: number | BigNumber;
-  height?: number;
-  callerNonce?: number;
-  log?: any;
-  returnType?: ContractCallReturnType;
-}
-
-interface Contract {
-  abiVersion: number;
-  active: boolean;
-  deposit: number | BigNumber;
-  id: string;
-  ownerId: string;
-  referrerIds: string[];
-  vmVersion: number;
-}
 
 /**
  * Channel
@@ -172,7 +147,10 @@ export default class Channel {
    * @param options.sign - Function which verifies and signs transactions
    */
   static async initialize(options: ChannelOptions): Promise<Channel> {
-    const channel = new Channel();
+    return Channel._initialize(new Channel(), options);
+  }
+
+  static async _initialize<T extends Channel>(channel: T, options: ChannelOptions): Promise<T> {
     await initialize(
       channel,
       options.existingFsmId != null ? handlers.awaitingReconnection : handlers.awaitingConnection,
@@ -275,7 +253,7 @@ export default class Channel {
     return this._fsmId;
   }
 
-  async #enqueueAction(
+  protected async enqueueAction(
     action: () => { handler: ChannelHandler; state?: Partial<ChannelState> },
   ): Promise<any> {
     return enqueueAction(
@@ -326,7 +304,7 @@ export default class Channel {
       errorCode?: number;
       errorMessage?: string;
     }> {
-    return this.#enqueueAction(() => {
+    return this.enqueueAction(() => {
       notify(this, 'channels.update.new', {
         from, to, amount, meta: metadata,
       });
@@ -420,7 +398,7 @@ export default class Channel {
    * ```
    */
   async leave(): Promise<{ channelId: Encoded.Bytearray; signedTx: Encoded.Transaction }> {
-    return this.#enqueueAction(() => {
+    return this.enqueueAction(() => {
       notify(this, 'channels.leave');
       return { handler: handlers.awaitingLeave };
     });
@@ -441,7 +419,7 @@ export default class Channel {
    * ```
    */
   async shutdown(sign: SignTx): Promise<Encoded.Transaction> {
-    return this.#enqueueAction(() => {
+    return this.enqueueAction(() => {
       notify(this, 'channels.shutdown');
       return {
         handler: handlers.awaitingShutdownTx,
@@ -505,7 +483,7 @@ export default class Channel {
     { onOnChainTx, onOwnWithdrawLocked, onWithdrawLocked }:
     Pick<ChannelState, 'onOnChainTx' | 'onOwnWithdrawLocked' | 'onWithdrawLocked'> = {},
   ): Promise<{ accepted: boolean; signedTx: Encoded.Transaction }> {
-    return this.#enqueueAction(() => {
+    return this.enqueueAction(() => {
       notify(this, 'channels.withdraw', { amount });
       return {
         handler: handlers.awaitingWithdrawTx,
@@ -575,7 +553,7 @@ export default class Channel {
     { onOnChainTx, onOwnDepositLocked, onDepositLocked }:
     Pick<ChannelState, 'onOnChainTx' | 'onOwnDepositLocked' | 'onDepositLocked'> = {},
   ): Promise<{ accepted: boolean; state: ChannelState }> {
-    return this.#enqueueAction(() => {
+    return this.enqueueAction(() => {
       notify(this, 'channels.deposit', { amount });
       return {
         handler: handlers.awaitingDepositTx,
@@ -586,319 +564,6 @@ export default class Channel {
           onDepositLocked,
         },
       };
-    });
-  }
-
-  /**
-   * Trigger create contract update
-   *
-   * The create contract update is creating a contract inside the channel's internal state tree.
-   * The update is a change to be applied on top of the latest state.
-   *
-   * That would create a contract with the poster being the owner of it. Poster commits initially
-   * a deposit amount of coins to the new contract.
-   *
-   * @param options - Options
-   * @param options.code - Api encoded compiled AEVM byte code
-   * @param options.callData - Api encoded compiled AEVM call data for the code
-   * @param options.deposit - Initial amount the owner of the contract commits to it
-   * @param options.vmVersion - Version of the Virtual Machine
-   * @param options.abiVersion - Version of the Application Binary Interface
-   * @param sign - Function which verifies and signs create contract transaction
-   * @example
-   * ```js
-   * channel.createContract({
-   *   code: 'cb_HKtpipK4aCgYb17wZ...',
-   *   callData: 'cb_1111111111111111...',
-   *   deposit: 10,
-   *   vmVersion: 3,
-   *   abiVersion: 1
-   * }).then(({ accepted, signedTx, address }) => {
-   *   if (accepted) {
-   *     console.log('New contract has been created')
-   *     console.log('Contract address:', address)
-   *   } else {
-   *     console.log('New contract has been rejected')
-   *   }
-   * })
-   * ```
-   */
-  async createContract(
-    {
-      code, callData, deposit, vmVersion, abiVersion,
-    }: {
-      code: Encoded.ContractBytearray;
-      callData: Encoded.ContractBytearray;
-      deposit: number | BigNumber;
-      vmVersion: number;
-      abiVersion: number;
-    },
-    sign: SignTx,
-  ): Promise<{
-      accepted: boolean; signedTx: Encoded.Transaction; address: Encoded.ContractAddress;
-    }> {
-    return this.#enqueueAction(() => {
-      notify(this, 'channels.update.new_contract', {
-        code,
-        call_data: callData,
-        deposit,
-        vm_version: vmVersion,
-        abi_version: abiVersion,
-      });
-      return {
-        handler: handlers.awaitingNewContractTx,
-        state: { sign },
-      };
-    });
-  }
-
-  /**
-   * Trigger call a contract update
-   *
-   * The call contract update is calling a preexisting contract inside the channel's
-   * internal state tree. The update is a change to be applied on top of the latest state.
-   *
-   * That would call a contract with the poster being the caller of it. Poster commits
-   * an amount of coins to the contract.
-   *
-   * The call would also create a call object inside the channel state tree. It contains
-   * the result of the contract call.
-   *
-   * It is worth mentioning that the gas is not consumed, because this is an off-chain
-   * contract call. It would be consumed if it were an on-chain one. This could happen
-   * if a call with a similar computation amount is to be forced on-chain.
-   *
-   * @param options - Options
-   * @param options.amount - Amount the caller of the contract commits to it
-   * @param options.callData - ABI encoded compiled AEVM call data for the code
-   * @param options.contract - Address of the contract to call
-   * @param options.abiVersion - Version of the ABI
-   * @param sign - Function which verifies and signs contract call transaction
-   * @example
-   * ```js
-   * channel.callContract({
-   *   contract: 'ct_9sRA9AVE4BYTAkh5RNfJYmwQe1NZ4MErasQLXZkFWG43TPBqa',
-   *   callData: 'cb_1111111111111111...',
-   *   amount: 0,
-   *   abiVersion: 1
-   * }).then(({ accepted, signedTx }) => {
-   *   if (accepted) {
-   *     console.log('Contract called succesfully')
-   *   } else {
-   *     console.log('Contract call has been rejected')
-   *   }
-   * })
-   * ```
-   */
-  async callContract(
-    {
-      amount, callData, contract, abiVersion,
-    }: CallContractOptions,
-    sign: SignTx,
-  ): Promise<{ accepted: boolean; signedTx: Encoded.Transaction }> {
-    return this.#enqueueAction(() => {
-      notify(this, 'channels.update.call_contract', {
-        amount,
-        call_data: callData,
-        contract_id: contract,
-        abi_version: abiVersion,
-      });
-      return {
-        handler: handlers.awaitingCallContractUpdateTx,
-        state: { sign },
-      };
-    });
-  }
-
-  /**
-   * Trigger a force progress contract call
-   * This call is going on-chain
-   * @param options - Options
-   * @param options.amount - Amount the caller of the contract commits to it
-   * @param options.callData - ABI encoded compiled AEVM call data for the code
-   * @param options.contract - Address of the contract to call
-   * @param options.abiVersion - Version of the ABI
-   * @param options.gasPrice=1000000000]
-   * @param options.gasLimit=1000000]
-   * @param sign - Function which verifies and signs contract force progress transaction
-   * @param callbacks - Callbacks
-   * @example
-   * ```js
-   * channel.forceProgress({
-   *   contract: 'ct_9sRA9AVE4BYTAkh5RNfJYmwQe1NZ4MErasQLXZkFWG43TPBqa',
-   *   callData: 'cb_1111111111111111...',
-   *   amount: 0,
-   *   abiVersion: 1,
-   *   gasPrice: 1000005554
-   * }).then(({ accepted, signedTx }) => {
-   *   if (accepted) {
-   *     console.log('Contract force progress call successful')
-   *   } else {
-   *     console.log('Contract force progress call has been rejected')
-   *   }
-   * })
-   * ```
-   */
-  async forceProgress(
-    {
-      amount, callData, contract, abiVersion, gasLimit = 1000000, gasPrice = MIN_GAS_PRICE,
-    }: {
-      amount: number;
-      callData: Encoded.ContractBytearray;
-      contract: Encoded.ContractAddress;
-      abiVersion: number;
-      gasLimit?: number;
-      gasPrice?: number;
-    },
-    sign: SignTx,
-    { onOnChainTx }: Pick<ChannelState, 'onOnChainTx'> = {},
-  ): Promise<{
-      accepted: boolean;
-      signedTx: Encoded.Transaction;
-      tx: Encoded.Transaction | Uint8Array;
-    }> {
-    return this.#enqueueAction(() => {
-      notify(this, 'channels.force_progress', {
-        amount,
-        call_data: callData,
-        contract_id: contract,
-        abi_version: abiVersion,
-        gas_price: gasPrice,
-        gas: gasLimit,
-      });
-      return {
-        handler: handlers.awaitingCallContractForceProgressUpdate,
-        state: { sign, onOnChainTx },
-      };
-    });
-  }
-
-  /**
-   * Call contract using dry-run
-   *
-   * In order to get the result of a potential contract call, one might need to
-   * dry-run a contract call. It takes the exact same arguments as a call would
-   * and returns the call object.
-   *
-   * The call is executed in the channel's state, but it does not impact the state
-   * whatsoever. It uses as an environment the latest channel's state and the current
-   * top of the blockchain as seen by the node.
-   *
-   * @param options - Options
-   * @param options.amount - Amount the caller of the contract commits to it
-   * @param options.callData - ABI encoded compiled AEVM call data for the code
-   * @param options.contract - Address of the contract to call
-   * @param options.abiVersion - Version of the ABI
-   * @example
-   * ```js
-   * channel.callContractStatic({
-   *   contract: 'ct_9sRA9AVE4BYTAkh5RNfJYmwQe1NZ4MErasQLXZkFWG43TPBqa',
-   *   callData: 'cb_1111111111111111...',
-   *   amount: 0,
-   *   abiVersion: 1
-   * }).then(({ returnValue, gasUsed }) => {
-   *   console.log('Returned value:', returnValue)
-   *   console.log('Gas used:', gasUsed)
-   * })
-   * ```
-   */
-  async callContractStatic(
-    {
-      amount, callData, contract, abiVersion,
-    }: {
-      amount: number;
-      callData: Encoded.ContractBytearray;
-      contract: Encoded.ContractAddress;
-      abiVersion: number;
-    },
-  ): Promise<CallContractOptions> {
-    return snakeToPascalObjKeys(await call(this, 'channels.dry_run.call_contract', {
-      amount,
-      call_data: callData,
-      contract_id: contract,
-      abi_version: abiVersion,
-    }));
-  }
-
-  /**
-   * Get contract call result
-   *
-   * The combination of a caller, contract and a round of execution determines the
-   * contract call. Providing an incorrect set of those results in an error response.
-   *
-   * @param options - Options
-   * @param options.caller - Address of contract caller
-   * @param options.contract - Address of the contract
-   * @param options.round - Round when contract was called
-   * @example
-   * ```js
-   * channel.getContractCall({
-   *   caller: 'ak_Y1NRjHuoc3CGMYMvCmdHSBpJsMDR6Ra2t5zjhRcbtMeXXLpLH',
-   *   contract: 'ct_9sRA9AVE4BYTAkh5RNfJYmwQe1NZ4MErasQLXZkFWG43TPBqa',
-   *   round: 3
-   * }).then(({ returnType, returnValue }) => {
-   *   if (returnType === 'ok') console.log(returnValue)
-   * })
-   * ```
-   */
-  async getContractCall(
-    { caller, contract, round }: {
-      caller: Encoded.AccountAddress;
-      contract: Encoded.ContractAddress;
-      round: number;
-    },
-  ): Promise<{
-      returnType: ContractCallReturnType;
-      returnValue: Encoded.ContractBytearray;
-      gasPrice: number | BigNumber;
-      gasUsed: number | BigNumber;
-      height: number;
-      log: string;
-    }> {
-    return snakeToPascalObjKeys(
-      await call(this, 'channels.get.contract_call', {
-        caller_id: caller,
-        contract_id: contract,
-        round,
-      }),
-    );
-  }
-
-  /**
-   * Get the latest contract state
-   *
-   * @param contract - Address of the contract
-   * @example
-   * ```js
-   * channel.getContractState(
-   *   'ct_9sRA9AVE4BYTAkh5RNfJYmwQe1NZ4MErasQLXZkFWG43TPBqa'
-   * ).then(({ contract }) => {
-   *   console.log('deposit:', contract.deposit)
-   * })
-   * ```
-   */
-  async getContractState(
-    contract: Encoded.ContractAddress,
-  ): Promise<{ contract: Contract; contractState: object }> {
-    const result = await call(this, 'channels.get.contract', { pubkey: contract });
-    return snakeToPascalObjKeys({
-      ...result,
-      contract: snakeToPascalObjKeys(result.contract),
-    });
-  }
-
-  /**
-   * Clean up all locally stored contract calls
-   *
-   * Contract calls are kept locally in order for the participant to be able to look them up.
-   * They consume memory and in order for the participant to free it - one can prune all messages.
-   * This cleans up all locally stored contract calls and those will no longer be available for
-   * fetching and inspection.
-   */
-  async cleanContractCalls(): Promise<void> {
-    return this.#enqueueAction(() => {
-      notify(this, 'channels.clean_contract_calls');
-      return { handler: handlers.awaitingCallsPruned };
     });
   }
